refactor(ECharts component): enable to define the chart options from outside
