+++ src/components/apache-echarts/ECharts/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ECharts.svelte';
+
+export type { EChartsType, EChartsOption } from './types';
