+++ src/components/apache-echarts/ECharts/types.ts
@@ -1,0 +1,11 @@
+import type { EChartsType, EChartsOption } from 'echarts';
+
+export interface EChartsProps {
+  // TODO: remove
+  message?: string | undefined;
+
+  chart?: EChartsType | undefined;
+  options?: EChartsOption | undefined;
+}
+
+export type { EChartsType, EChartsOption } from 'echarts';
