--- src/components/apache-echarts/base/types.ts
@@ -1,4 +1,0 @@
-export interface EChartsProps {
-  // TODO: define
-  message?: string;
-}
