+++ src/components/apache-echarts/examples/StarterExample.svelte
@@ -1,5 +1,26 @@
 <script lang="ts">
-  import ECharts from '~/components/apache-echarts/base/ECharts.svelte';
+  import ECharts, {
+    type EChartsOption
+  } from '~/components/apache-echarts/ECharts';
+
+  const options: EChartsOption = {
+    title: {
+      text: 'Getting Started',
+      subtext: 'ECharts Basic Example'
+    },
+    tooltip: {},
+    xAxis: {
+      data: ['shirt', 'cardigan', 'chiffon', 'pants', 'heels', 'socks']
+    },
+    yAxis: {},
+    series: [
+      {
+        name: 'sales',
+        type: 'bar',
+        data: [5, 20, 36, 10, 10, 20]
+      }
+    ]
+  };
 </script>
 
-<ECharts>Loading Starter Example Chart...</ECharts>
+<ECharts {options}>Loading Starter Example Chart...</ECharts>
